% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{cellscapeOutput}
\alias{cellscapeOutput}
\title{Widget output function for use in Shiny}
\usage{
cellscapeOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

