% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvTree.R
\name{cnvTreeOutput}
\alias{cnvTreeOutput}
\title{Widget output function for use in Shiny}
\usage{
cnvTreeOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

