% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{findMode}
\alias{findMode}
\title{function to find the mode of a vector}
\usage{
findMode(x)
}
\arguments{
\item{x}{-- vector of numbers}
}
\description{
function to find the mode of a vector
}
\examples{
findMode(c(1,1,19,1))
}

