% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvTree.R
\name{getChromBoxInfo}
\alias{getChromBoxInfo}
\title{function to get chromosome box pixel info}
\usage{
getChromBoxInfo(chrom_bounds, n_bp_per_pixel)
}
\arguments{
\item{{DataFrame}}{chrom_bounds -- chromosome boundaries}

\item{{Integer}}{n_bp_per_pixel -- number of base pairs per pixel}
}
\description{
function to get chromosome box pixel info
}

