% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getCNVHeatmapForEachSC}
\alias{getCNVHeatmapForEachSC}
\title{function to get information (chr, start, end, mode_cnv) for each pixel}
\usage{
getCNVHeatmapForEachSC(cnv_data, chrom_bounds, n_bp_per_pixel)
}
\arguments{
\item{cnv_data}{-- data frame of copy number variant segments data}

\item{chrom_bounds}{-- data frame of chromosome boundaries}

\item{n_bp_per_pixel}{-- integer of number of base pairs per pixel}
}
\description{
function to get information (chr, start, end, mode_cnv) for each pixel
}

