% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getEmptyGrid}
\alias{getEmptyGrid}
\title{Function to get data frame of pixels}
\usage{
getEmptyGrid(hm_sc_ids_ordered, ncols)
}
\arguments{
\item{hm_sc_ids_ordered}{-- array of single cell ids in order}

\item{ncols}{-- number of columns in heatmap/grid}
}
\description{
Function to get data frame of pixels
}

