% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvTree.R
\name{getGenomeLength}
\alias{getGenomeLength}
\title{function to get the genome length}
\usage{
getGenomeLength(chrom_bounds)
}
\arguments{
\item{{Data}{Frame} chrom_bounds -- chromosome boundaries}
}
\description{
function to get the genome length
}

