% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getMutOrder}
\alias{getMutOrder}
\title{function to get mutation order for targeted data}
\usage{
getMutOrder(mut_data)
}
\arguments{
\item{mut_data}{-- data frame of mutations data}
}
\description{
function to get mutation order for targeted data
}

