% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvTree.R
\name{getNBPPerPixel}
\alias{getNBPPerPixel}
\title{function to get the number of base pairs per pixel}
\usage{
getNBPPerPixel(ncols, chrom_bounds, genome_length)
}
\arguments{
\item{{Integer}}{ncols -- number of columns (pixels) to fill}

\item{{Data}{Frame} chrom_bounds -- chromosome boundaries}

\item{{Integer}}{genome_length -- length of the genome}
}
\description{
function to get the number of base pairs per pixel
}

