% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getTargetedHeatmapForEachSC}
\alias{getTargetedHeatmapForEachSC}
\title{function to get targeted heatmap information}
\usage{
getTargetedHeatmapForEachSC(mut_data, mut_order, heatmapWidth)
}
\arguments{
\item{mut_data}{-- data frame of mutations data}

\item{mut_order}{-- array of order of mutations for heatmap (chromosome:coordinate)}

\item{heatmapWidth}{-- number for width of the heatmap (in pixels)}
}
\description{
function to get targeted heatmap information
}

