% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvTree.R
\name{renderCnvTree}
\alias{renderCnvTree}
\title{Widget render function for use in Shiny}
\usage{
renderCnvTree(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

