% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{cellscape}
\alias{cellscape}
\title{CellScape}
\usage{
cellscape(cnv_data = NULL, mut_data = NULL, mut_order = NULL, tree_edges,
  gtype_tree_edges = NULL, sc_annot = NULL, clone_colours = "NA",
  timepoint_title = "Timepoint", clone_title = "Clone",
  xaxis_title = "Time Point", yaxis_title = "Clonal Prevalence",
  phylogeny_title = "Clonal Phylogeny", value_type = NULL,
  node_type = "Cell", display_node_ids = FALSE,
  prop_of_clone_threshold = 0.2, vaf_threshold = 0.05,
  show_warnings = TRUE, width = 900, height = 800)
}
\arguments{
\item{cnv_data}{{Data frame} (Required if mut_data not provided) Single cell copy number segments data. 
Note that every single cell id must be present in the tree_edges data frame.
Format: columns are (1) {String} "single_cell_id" - single cell id
                    (2) {String} "chr" - chromosome number
                    (3) {Number} "start" - start position
                    (4) {Number} "end" - end position
                    (5) {Number} "integer_copy_number" - copy number state.}

\item{mut_data}{{Data frame} (Required if cnv_data not provided) Single cell targeted mutation data frame.
Note that every single cell id must be present in the tree_edges data frame.
Format: columns are (1) {String} "single_cell_id" - single cell id
                    (2) {String} "chr" - chromosome number
                    (3) {Number} "coord" - genomic coordinate
                    (5) {Number} "VAF" - variant allele frequency [0, 1].}

\item{mut_order}{{Array} (Optional) Mutation order for targeted mutation heatmap 
(each mutation should consist of a string in the form "chrom:coord"). 
Default will use a clustering function to determine mutation order.}

\item{tree_edges}{{Data frame} Edges for the single cell phylogenetic tree.
Format: columns are (1) {String} "source" - edge source (single cell id)
                    (2) {String} "target" - edge target (single cell id)
                    (2) {Number} (Optional) "dist" - edge distance}

\item{gtype_tree_edges}{{DataFrame} (Required for TimeScape) Genotype tree edges of a rooted tree.
Format: columns are (1) {String} "source" - source node id
                    (2) {String} "target" - target node id.}

\item{sc_annot}{{Data frame} (Required for TimeScape) Annotations (genotype and sample id) for each single cell.
Format: columns are (1) {String} "single_cell_id" - single cell id
                    (2) {String} "genotype" - genotype assignment
                    (3) {String} (Optional) "timepoint" - id of the sampled time point. 
                                               Note: in the case of time points, they will be ordered alphabetically}

\item{clone_colours}{{DataFrame} (Optional) Clone ids and their corresponding colours 
Format: columns are (1) {String} "clone_id" - the clone ids
                    (2) {String} "colour" - the corresponding Hex colour for each clone id.}

\item{timepoint_title}{{String} (Optional) Legend title for timepoint groups. Default is "Timepoint".}

\item{clone_title}{{String} (Optional) Legend title for clones. Default is "Clone".}

\item{xaxis_title}{{String} (Optional) For TimeScape - x-axis title. Default is "Time Point".}

\item{yaxis_title}{{String} (Optional) For TimeScape - y-axis title. Default is "Clonal Prevalence".}

\item{phylogeny_title}{{String} (Optional) For TimeScape - legend phylogeny title. Default is "Clonal Phylogeny".}

\item{value_type}{{String} (Optional) The type of value plotted in heatmap - will affect legend and heatmap tooltips. 
Default is "VAF" for mutation data, and "CNV" for copy number data.}

\item{node_type}{{String} (Optional) The type of node plotted in single cell phylogeny - will affect phylogeny tooltips. 
Default is "Cell".}

\item{display_node_ids}{{Boolean} (Optional) Whether or not to display the single cell ID within the tree nodes. Default is FALSE.}

\item{prop_of_clone_threshold}{{Number} (Optional) Used for the ordering of targeted mutations. The minimum proportion of a clone 
to have a mutation in order to consider the mutation as present within that clone. Default is 0.1.}

\item{vaf_threshold}{{Number} (Optional) Used for the ordering of targeted mutations. The minimum variant allele frequency for a mutation to
be considered as present within a single cell. Default is 0.05.}

\item{show_warnings}{{Boolean} (Optional) Whether or not to show any warnings. Default is TRUE.}

\item{width}{{Number} (Optional) Width of the plot.}

\item{height}{{Number} (Optional) Height of the plot.}
}
\description{
\code{cellscape} explores single cell copy number profiles in the context of a single cell phylogeny.
}
\examples{

library("cellscape")

# EXAMPLE 1 - TARGETED MUTATION DATA

# single cell tree edges
tree_edges <- read.csv(system.file("extdata", "targeted_tree_edges.csv", package = "cellscape"))
# targeted mutations
targeted_data <- read.csv(system.file("extdata", "targeted_muts.csv", package = "cellscape"))
# genotype tree edges
gtype_tree_edges <- data.frame("source"=c("Ancestral", "Ancestral", "B","C", "D"), "target"=c("A", "B", "C", "D", "E"))
# annotations
sc_annot <- read.csv(system.file("extdata", "targeted_annots.csv", package = "cellscape"))
# mutation order
mut_order <- scan(system.file("extdata", "targeted_mut_order.txt", package = "cellscape"), what=character())
# run cellscape
cellscape(mut_data=targeted_data, tree_edges=tree_edges, sc_annot = sc_annot, gtype_tree_edges=gtype_tree_edges, mut_order=mut_order)

# EXAMPLE 2 - COPY NUMBER DATA

# single cell tree edges
tree_edges <- read.csv(system.file("extdata", "cnv_tree_edges.csv", package = "cellscape"))
# cnv segments data
cnv_data <- read.csv(system.file("extdata", "cnv_data.csv", package = "cellscape"))
# annotations
sc_annot <- read.csv(system.file("extdata", "cnv_annots.tsv", package = "cellscape"), sep="\\t")
# custom clone colours
clone_colours <- data.frame( clone_id = c("1","2","3"), colour = c("7fc97f", "beaed4", "fdc086"))
# run cellscape
cellscape(cnv_data=cnv_data, tree_edges=tree_edges, sc_annot=sc_annot, width=800, height=475, show_warnings=FALSE, clone_colours=clone_colours)
}

