% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{cellscapeOutput}
\alias{cellscapeOutput}
\title{Widget output function for use in Shiny}
\usage{
cellscapeOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{-- id of output}

\item{width}{-- width of output}

\item{height}{-- height of output}
}
\description{
Widget output function for use in Shiny
}

