% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{dfs_tree}
\alias{dfs_tree}
\title{Get depth first search of a tree}
\usage{
dfs_tree(edges, cur_root, dfs_arr)
}
\arguments{
\item{edges}{-- edges of tree}

\item{cur_root}{-- current root of the tree}

\item{dfs_arr}{-- array of depth first search results to be filled}
}
\description{
Get depth first search of a tree
}
\examples{
dfs_tree(data.frame(source = c("1","1","2","2","5","6"), target=c("2","5","3","4","6","7")), "1", c())
}

