% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getChromBounds}
\alias{getChromBounds}
\title{function to get min and max values for each chromosome}
\usage{
getChromBounds(chroms, cnv_data)
}
\arguments{
\item{chroms}{-- vector of chromosome names}

\item{cnv_data}{-- copy number data}
}
\description{
function to get min and max values for each chromosome
}

