% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getChromBoxInfo}
\alias{getChromBoxInfo}
\title{function to get chromosome box pixel info}
\usage{
getChromBoxInfo(chrom_bounds, n_bp_per_pixel)
}
\arguments{
\item{chrom_bounds}{-- data frame of chromosome boundaries}

\item{n_bp_per_pixel}{-- integer of number of base pairs per pixel}
}
\description{
function to get chromosome box pixel info
}

