% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvTree.R
\name{getCNVHeatmapForEachSC}
\alias{getCNVHeatmapForEachSC}
\title{function to get information (chr, start, end, mode_cnv) for each pixel}
\usage{
getCNVHeatmapForEachSC(cnv_data, chrom_bounds, n_bp_per_pixel)
}
\arguments{
\item{{Data}{Frame} cnv_data -- copy number variant segments data}

\item{{Data}{Frame} chrom_bounds -- chromosome boundaries}

\item{{Integer}}{n_bp_per_pixel -- number of base pairs per pixel}
}
\description{
function to get information (chr, start, end, mode_cnv) for each pixel
}

