% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getGenomeLength}
\alias{getGenomeLength}
\title{function to get the genome length}
\usage{
getGenomeLength(chrom_bounds)
}
\arguments{
\item{chrom_bounds}{-- data frame of chromosome boundaries}
}
\description{
function to get the genome length
}

