% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getNBPPerPixel}
\alias{getNBPPerPixel}
\title{function to get the number of base pairs per pixel}
\usage{
getNBPPerPixel(ncols, chrom_bounds, genome_length)
}
\arguments{
\item{ncols}{--  integer of number of columns (pixels) to fill}

\item{chrom_bounds}{--  data frame of chromosome boundaries}

\item{genome_length}{--  integer of length of the genome}
}
\description{
function to get the number of base pairs per pixel
}

