% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvTree.R
\name{getTargetedHeatmapForEachSC}
\alias{getTargetedHeatmapForEachSC}
\title{function to get targeted heatmap information}
\usage{
getTargetedHeatmapForEachSC(mut_data, heatmapWidth)
}
\arguments{
\item{{Data}{Frame} mut_data -- mutations data}

\item{{Number}}{heatmapWidth -- width of the heatmap (in pixels)}
}
\description{
function to get targeted heatmap information
}

