% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellscape.R
\name{getTargetedHeatmapForEachSC}
\alias{getTargetedHeatmapForEachSC}
\title{function to get targeted heatmap information}
\usage{
getTargetedHeatmapForEachSC(mut_data, mut_order, heatmapWidth)
}
\arguments{
\item{{DataFrame}}{mut_data -- mutations data}

\item{{Array}}{mut_order -- order of mutations for heatmap (chromosome:coordinate)}

\item{{Number}}{heatmapWidth -- width of the heatmap (in pixels)}
}
\description{
function to get targeted heatmap information
}

